Moved unit tests into their own file and imported form json
